+++ examples/MemoryBenchmark/generate_readme.py
@@ -27,7 +27,7 @@ memory and static RAM sizes were recorded. The `FEATURE_BASELINE` selection is
 the baseline, and its memory usage  numbers are subtracted from the subsequent
 `FEATURE_*` memory usage.
 
-**Version**: AceTime v2.1.0
+**Version**: AceTime v2.1.1
 
 **DO NOT EDIT**: This file was auto-generated using `make README.md`.